backend/database.py
# input:  [Environment variables, SQLAlchemy engine/session/base]
# output: [Database engine, session factory, declarative base, and FK pragma hook]
# pos:    [Database bootstrap and connection configuration]
#
# ⚠️ When this file is updated:
#    1. Update these header comments
#    2. Update the INDEX.md of the folder this file belongs to

from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
from sqlalchemy import event
from dotenv import load_dotenv
from pathlib import Path
import os

BASE_DIR = Path(__file__).resolve().parent
ENVIRONMENT = os.getenv("ENVIRONMENT", "development")

# Load local .env for development so DB_PATH can be configured without code changes.
if ENVIRONMENT == "development":
    env_local_path = BASE_DIR / ".env"
    if env_local_path.exists():
        load_dotenv(env_local_path)

DATABASE_URL = os.getenv("DATABASE_URL")
DB_PATH = os.getenv("DB_PATH")

if DATABASE_URL:
    SQLITE_URL = DATABASE_URL
else:
    if DB_PATH:
        db_path = Path(DB_PATH)
        if not db_path.is_absolute():
            db_path = BASE_DIR / db_path
    else:
        db_path = BASE_DIR / "semestra.db"
    DB_PATH = str(db_path)
    SQLITE_URL = f"sqlite:///{DB_PATH}"

engine = create_engine(
    SQLITE_URL, connect_args={"check_same_thread": False}
)

@event.listens_for(engine, "connect")
def _set_sqlite_pragma(dbapi_connection, _connection_record):
    cursor = dbapi_connection.cursor()
    cursor.execute("PRAGMA foreign_keys=ON")
    cursor.close()

SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = declarative_base()

def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()
