backend/requirements.txt
# input:  [Python package index and project dependency constraints]
# output: [Backend runtime dependency list]
# pos:    [Environment dependency manifest for backend]
#
# ⚠️ When this file is updated:
#    1. Update these header comments
#    2. Update the INDEX.md of the folder this file belongs to

fastapi
uvicorn
sqlalchemy
pydantic
icalendar
python-multipart
python-jose[cryptography]
passlib[bcrypt]
requests
google-auth

python-dotenv
